requirements.txt
appdirs==1.4.4
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
binance-connector==1.18.0
binance-futures-connector==3.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
fonttools==4.38.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.2
mplfinance==0.12.9b5
multitasking==0.0.11
numpy==1.23.5
packaging==21.3
pandas==1.5.1
pandas-ta==0.3.14b0
Pillow==9.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
service-identity==21.1.0
six==1.16.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
yfinance==0.1.87
zope.interface==5.5.2
